backend/requirements.txt
pandas==2.0.0
numpy==1.24.3
scikit-learn==1.2.2
python-dotenv==1.0.0
supabase==1.0.3
requests==2.28.2
